Handle MultiPolygon Voronoi Cells in load_cell
A Voronoi cell may be discontiguous, for example if a tower also serves
an island off the mainland. These are stored in the JSON as
`MultiPolygon`s. To handle this, `load_cell` now accepts more of the
JSON so it can access the `type` key to see if the JSON is for a
`MultiPolygon` or a `Polygon`. Processing differs by type, but in any
event a `MultiPolygon` is returned. It will have only one member if the
JSON was for a `Polygon`.