Add Script and Function to Map Cells to Seeds
Since we are given the Voronoi seeds (cell towers) separately from the
Voronoi cells and without a mapping between them, the script computes
this mapping by finding which cell each seed is contained within. This
reveals that the mappings do not appear to be related to index, but that
there is a one-to-one correspondence between the provided seeds and
cells.