Keep import paths relative to script file
Import paths had been changed to be relative to the root of the
repository for easy importing by Sphinx, but this prevented the script
files from finding their dependencies. Instead, the import paths were
made relative to the script files (in the
`mobility_pipeline/mobility_pipeline` directory), and Sphinx was
configured to include that directiory in its search path.