Use Ratio for Checking Continuous, Disjoint Cells
Instead of specifying how much area we allow to deviate between the sum
of polygon areas and the area of the polygons' union in contiguous and
disjoint cells as a constant amount of area, require that the deviance
be below some fraction of the union area. This is preferable because
over larger areas, equally acceptable deviances will add up. Therefore,
a constant allowed deviance is inappropriate and will either be too
strict for large countries or too lenient for small countries. The
threshold was chosen arbitrarily but to work for Brazil.