Create Script to Run For Each Day's Data
It will be common for a user to process many day's worth of data for a
single country. In such a scenario, we only need to generate the overlap
matrices and other country-specific data once. This commit creates a
script to take in that stored data and generate a single day's worth of
mobility information.

See #25 for details.